Lands the full pointer-value migration (review architectural decision #1) and the
server request-parallelism it unblocks (#2).

- Value is now {hi uint64; ptr unsafe.Pointer}; every heap kind is a Go-GC-traced
  pointer. The ~1.8k-line custom collector, ID registries, KeepAlive, weakref
  liveness, block slot+gen recycling, and the AutoIDRegistry pressure apparatus
  are deleted.
- Server: the single-goroutine VMWorker funnel is retired for a RWMutex gate —
  read/eval requests run concurrently (DoConcurrent) on per-request interpreters
  (vm.RunIsolated), structural mutations run exclusively (Do). Schema-mutating
  Evaluate is routed to the exclusive gate (SourceMayMutateSchema/DoForSource).

Measured: HotPath geomean time -18% (alloc-heavy paths win big, tight scalar
dispatch pays the 16-byte copy); BenchmarkServerEvalParallel now scales
1.85x/3.21x/4.35x at 2/4/8 cores vs the pre-migration flat ~7.6ms/op.

Gates green: go test ./... , go test -race ./vm/ ./server/ (0 DATA RACE),
1166/1166 doctests. Post-merge review found and fixed a forked-interpreter
global-write regression (H1) and a HandleStore data race (L4).

+++ compiler/codegen.go
@@ -545,7 +545,7 @@ func (c *Compiler) compileDictionaryLiteral(dict *DictionaryLiteral) {
 // semantically equal, and registry-backed values (Strings/Symbols) are
 // interned at construction so equal payloads share an ID.
 func (c *Compiler) addLiteral(value vm.Value) int {
-	key := uint64(value)
+	key := value
 	if idx, ok := c.frame.literalMap[key]; ok {
 		return idx
 	}
